Refine code related to table and partition stats
